+++ SchoolDataImporter/Enums/FilterType.cs
@@ -1,0 +1,14 @@
+﻿namespace SchoolDataImporter.Enums
+{
+    public enum FilterType
+    {
+        Gender,
+        Status,
+        Grades,
+        BusRoutes,
+        Houses,
+        Hostels,
+        Staff,
+        GoverningBody
+    }
+}